requirements.txt
absl-py==0.7.1
astor==0.8.0
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.16.4
Pillow==6.0.0
protobuf==3.8.0
six==1.12.0
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
Werkzeug==0.15.4
